Fixed bug where theme files couldn't be dragged onto Marabu
Turns out there is inconsistent behavior between Mac OS and Windows on how -webkit-app-region: drag is handled. On Windows, this blocks window events (like 'dragover' and 'drop'), but does not on Mac. There is a workaround however, by adding -webkit-app-region: no-drag to a child-element of the element with drag, window events can be triggered on the child-element. This clearfixes the yu element giving it height, and adds no-drag to it.